requirement.txt
Django
django-filter
djangorestframework
PyJWT
django-cors-headers
djangorestframework-simplejwt
django_rest_passwordreset